Added requested changes to terms page
